requirements.txt
gunicorn
Flask
pandas
sklearn
numpy
matplotlib
PyYAML
